+++ jscomp/gentype_tests/typescript-react-example/src/Unboxed.gen.tsx
@@ -16,6 +16,14 @@ export type r2 = string;
 
 export type t = number[] | number | ((_1:number) => number);
 
+export type tabIndex = "0" | "1";
+
 export const testV1: (x:v1) => v1 = UnboxedBS.testV1;
 
 export const r2Test: (x:r2) => r2 = UnboxedBS.r2Test;
+
+export const a: tabIndex = UnboxedBS.a;
+
+export const b: tabIndex = UnboxedBS.b;
+
+export const zero: 0 = UnboxedBS.zero;