+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix issue with GenType and `result` introduced in rc.5. https://github.com/rescript-lang/rescript-compiler/pull/6464
 - Fix compiler crash when inlining complex constants in pattern matching. https://github.com/rescript-lang/rescript-compiler/pull/6471
 - Fix issue with generating async functions inside loops. https://github.com/rescript-lang/rescript-compiler/pull/6479
+- Fix issue with Gentype and string annotations with numbers such as `@as("0")`. https://github.com/rescript-lang/rescript-compiler/pull/6487
 
 
 # 11.0.0-rc.5